requirements.txt
huggingface_hub>=0.20.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
wandb>=0.16.0
python-dotenv>=1.0.0
